pyproject.toml
[project]
name = "fastai"
version = "0.0.0"
description = "A template for creating AI applications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.1",
    "argon2-cffi>=25.1.0",
    "asgi-correlation-id>=4.3.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "fastmcp>=2.5.2",
    "gunicorn>=23.0.0",
    "logfire[fastapi,sqlalchemy]>=3.16.1",
    "pwdlib[argon2]>=0.3.0",
    "authlib>=1.6.0",
    "joserfc>=1.6.0",
    "pydantic>=2.11.5",
    "pydantic-ai-slim[logfire,openai,openrouter,anthropic]>=0.2.12",
    "pydantic-settings>=2.9.1",
    "sqlalchemy[asyncio]>=2.0.41",
    "sqlmodel>=0.0.24",
    "starlette>=0.46.2",
    "structlog>=25.3.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "polylith-cli>=1.28.0",
    "pyright>=1.1.401",
    "ruff>=0.11.12",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
dev-mode-dirs = ["components", "bases", "development", "."]

[tool.polylith.bricks]
"bases/fastai/api" = "fastai/api"
"components/fastai/users" = "fastai/users"
"components/fastai/items" = "fastai/items"
"components/fastai/database" = "fastai/database"
"components/fastai/logger" = "fastai/logger"
"components/fastai/utils" = "fastai/utils"
"components/fastai/auth" = "fastai/auth"
"components/fastai/admin_v1" = "fastai/admin_v1"
"components/fastai/api_v1" = "fastai/api_v1"
"components/fastai/agents" = "fastai/agents"


[tool.pytest.ini_options]
markers = [
    "integration: tests that require a running database",
]

[tool.isort]
profile = "black"
known_first_party = ["fastai"]
line_length=88
